# documentation/swagger.yaml
swagger: "2.0"
# Cabecera del documento.
info:
  title: Delilah Restó
  description: Tercer proyecto de la carrera DWFS de Acamica
  version: 1.0.0
# Fin cabecera del documento.

#asi todas las rutas quedan https://DelilahResto.com/v2/users
host: "DelilahResto.com"
basePath: "/v2"

#tags son los temas sobre los que hago las rutas
tags:
- name: "Users"
  description: "Operations avalible on users"
- name: "Products"
  description: "Operations avalible on products"
- name: "Orders"
  description: "Operations avalible on orders"
  
schemes:
- "https"
- "http"

#define the MIME types supported by the API (consumes only operations with a request body, like post) If written in an operation level this will be overriden
consumes:
  - "application/json"
  - "application/xml"
produces:
  - "application/xml"
  - "application/json"

#rutas
paths:
  /user/register:
    post:
      #con tags identifico dentro de que categoria de tags de linea 10 escribo estas rutas
      tags:
      - "Users"
      summary: "Add a new user to the restó"
      description: "The new user complets the required information and is then added to the data base "
      operationId: "registerUser"
      parameters:
      - in: "body"
        name: "Needed information"
        description: "User object that needs to be added to the database"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        201:
          description: "User correctly added to database"
        401:
          description: "Username already exists, please pick another"
        500:
          description: "Internal Error"
  /user/logIn:
    post:
      tags:
      - "Users"
      summary: "Logs user into the system"
      description: "Logs user into the system"
      operationId: "loginUser"
      parameters:
      - in: "body"
        name: "Needed information"
        description: "Username and password to login"
        required: true
        schema:
          $ref: "#/definitions/Login"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid username/password supplied"
        500:
          description: "Internal Error"
  /Products:
    get:
      tags:
      - "Products"
      summary: "Returns all products"
      description: "Returns a table with all product related information"
      operationId: "getProducts"
      parameters: 
      - name: "Token"
        description: "token"
        in: "header"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedBothError'
  /Products/create:
    post:
      tags:
      - "Products"
      summary: "Adds a product to database"
      description: "This can only be done by the admin"
      operationId: "addProduct"
      parameters:
      - name: "Token"
        description: "Authorization: token"
        in: "header"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "product object that needs to be added to the database"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        201:
          description: "Product correctly added to database"
        401:
          description: "Product already exists"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedUserError'
  /products/deactivate/{Id}:
    delete:
      tags:
      - "Products"
      summary: "Indicate a product is not active in database"
      description: "This can only be done by the admin"
      operationId: "deactivateProduct"
      parameters:
      - name: "Token"
        in: "header"
        description: "Authorization: token"
        required: true
        type: "string"
      - name: "Id"
        in: "path"
        description: "Product id to deactivate"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Success"
        404:
          description: "Product not found"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedUserError'
  /products/{Id}:
    put:
      tags:
      - "Products"
      summary: "Update an existing product"
      description: "Change one or more properties of an existing product. Only admins can do it"
      operationId: "updateProduct"
      parameters:
      - name: "Token"
        description: "Authorization: token"
        in: "header"
        required: true
        type: "string"
      - name: "Id"
        in: "path"
        description: "Product id to modify"
        required: true
        type: "integer"
        format: "int64"
      - name: "body"
        in: "body"
        description: "Product object that is going to modify the current one"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        200:
          description: "Success"
        404:
          description: "Product not found"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedUserError'
  /orders:
    get:
      tags:
      - "Orders"
      summary: "Returns all orders (This can only be done by a admin) or all of the user orders (this can only be done by a user)"
      description: "Returns a table with all order related information"
      operationId: "getOrders"
      parameters: 
      - name: "Token"
        description: "Authorization: token"
        in: "header"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedBothError'
  /orders/create:
    post:
      tags:
      - "Orders"
      summary: "Add a new order"
      description: "Making an order. This can only be done by a user"
      operationId: "postOrder"
      parameters:
      - name: "Token"
        description: "Authorization: token"
        in: "header"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Order object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Order"
      responses:
        200:
          description: "Order created successfully"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedAdminError'
  /orders/{Id}:  
    patch:
      tags:
      - "Orders"
      summary: "Find purchase order by ID, to update status"
      description: "An administrator can update an order."
      operationId: "putOrderById"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of order that needs to be fetched"
        required: true
        type: "integer"
        maximum: 10.0
        minimum: 1.0
        format: "int64"
      - name: "New State"
        in: "body"
        description: "New state of the order"
        required: true
        schema:
          $ref: "#/definitions/NewState"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Order"
        500:
          description: "Internal Error"
        403:
          $ref: '#/responses/UnauthorizedUserError'

definitions:
  User:
    type: "object"
    properties:
      Username:
        type: "string"
        example: VickyCalvo
      Fullname:
        type: "string"
        example: Victoria Calvo
      Email:
        type: "string"
        example: victoriacalvo97@hotmail.com
      Phone number:
        type: "string"
        example: 3468408136
      Shipping adress:
        type: "string"
        example: De las acacias 139
      Password:
        type: "string"
        example: hola123
  Login:
    type: "object"
    properties:
      Username:
        type: "string"
        example: VickyCalvo
      Password:
        type: "string"
        example: hola123
  Product:
    type: "object"
    properties:
      Name:
        type: "string"
        example: Dulce de leche
      Descripción:
        type: "string"
        example: Marca la serenisima
      Price per unit:
        type: "number"
        example: 300
      ImageSrc:
        type: "string"
        example: "www.pinterest.com"
      isActive:
        type: "boolean"
        example: true
  Order:
    type: object
    properties:
      id_user:
        type: "integer"
        format: "int64"
        example: 1
      paymentMethod:
        type: "string"
        enum: ['efectivo', 'debito', 'credito']
        example: debito
      createdAt:
        type: "string"
        example: "2007-05-10 00:00:00"
      updatedAt:
        type: "string"
        example: "2007-05-10 00:00:00"
      paymentValue:
        type: "integer"
        example: 19.99
      state:
        type: "string"
        enum: ["entregado", "enviando", "preparando", "confirmado", "nuevo"]
        example: entregado
      products:
        type: "array"
        items:
          type: object
          properties:
            productId:
              type: "integer"
              example: 3
            productCuantity:
              type: "integer"
              example: 2
  NewState:
    type: object
    properties:
      state:
        type: "string"
        example: preprando/entregado/cancelado

   

#dentro de las responses de status hice referencia a estas formas de mostrar error   
responses:
  UnauthorizedAdminError:
    description: Administrators are not allowed to make this request
    headers:
      Authorization:
        description: token
        type: string
  UnauthorizedUserError:
    description: Users are not allowed to make this request
    headers:
      Authorization:
        description: token
        type: string
  UnauthorizedBothError:
    description: Nobody is allowed to make this request
    headers:
      Authorization:
        description: token
        type: string